+++ example/src/screens/MainStack/DevicenRFTestScreen/DevicenRFTestScreen.tsx
@@ -1,9 +1,9 @@
 import React, { useState, type Dispatch } from 'react'
-import type { TestStateType } from 'example/types'
 import type { NativeStackScreenProps } from '@react-navigation/native-stack'
 import { Device, type Base64 } from 'react-native-ble-plx'
 import { Platform, ScrollView } from 'react-native'
 import base64 from 'react-native-base64'
+import type { TestStateType } from '../../../types'
 import { BLEService } from '../../../services'
 import type { MainStackParamList } from '../../../navigation/navigators'
 import { AppButton, AppTextInput, ScreenDefaultContainer, TestStateDisplay } from '../../../components/atoms'
