+++ example/src/types/index.ts
